Create Metrique
This database table provides a structured storage solution for managing guest-related data in a standardized format. It allows for efficient retrieval, analysis, and manipulation of data related to property, guests, reservations, stay status, dates, source of business, market segments, rate types, room codes, and room types.
PropertyCode: Represents the code associated with a specific property.
SourceGuestID: Identifies the guest from the data source.
ReservationNumber: Stores the reservation number associated with a guest.
CustomerID: Represents the unique identifier of a customer.
StayStatus : Indicates the status of the guest's stay.
ArrivalDate: Stores the date of arrival for the guest.
DepartureDate: Stores the date of departure for the guest.
SourceOfBusiness: Represents the source or channel from which the business is obtained.
MarketSubSeg: Indicates the sub-segment of the market to which the guest belongs.
MarketSeg: Represents the market segment to which the guest belongs.
RateType : Stores the type of rate associated with the guest's reservation.
RoomCode: Represents the code associated with a specific room.
RoomTypeCode: Indicates the code associated with a specific room type.